Moved writer into own subpackage.
